Add sections view as alternative to vertical view (#19)
This adds a sections view of the app, where user navigates through sections which only show a subset of the total cards. As currently all cards were shown below each other, creating a long, vertical page, this brings a nicer way of a navigating through the app. The sections also group related cards giving the app a more intuitive feel